requirements.txt
pillow==11.1.0
qrcode==8.0
tkfontchooser==2.0.2
ttkthemes==3.2.2
